Apply default vanilla styles if none are explicitly specified
* Set vanillaStyles as default value in JsonFormsStyleContext
* Move vanillaStyles, StyleDef interface and the styles context to new module styles
* Adapt documentation
